some rough calibrations for prototype setup

+++ src/manual_move.py
@@ -3,4 +3,4 @@ from servo import CustomServo
 while True:
     i = input('enter cm to move: ')
     servo = CustomServo()
-    servo.move(cm=int(i))
+    servo.move(cm=float(i))
